ptrlist: specialize __add_ptr_list() for tag/notag
Currently, the function to add items to ptrlists, __add_ptr_list(),
always combine a 'raw' pointer and a tag.

But the vast majority of list entries have no tag and so just need
to store the raw pointer.

Avoid to do this operation when it's not needed by specializing
this function like this:
- __add_ptr_list() which directly store an untagged or 'cooked' ptr.
- __add_ptr_list_tag() which combine a pointer and a tag and then
  store this 'cooked' pointer by calling __add_ptr_list().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>