ptrlist: let all pointer lists have the same parametrized structure
Currently, all pointer lists have the same concrete structure,
struct ptr_list, which is also used as a generic pointer list.
But they're each declared with a phony type which just allows
to get back the type of the pointer:
	struct <name> { <TYPE> list[0]; };
vs.
	struct ptr_list {
		....
		void * list[NR];
	};

Since the declaration doesn't match the real representation
every operation done on pointer lists must cast forth & back the
list pointer to the generic type. OTOH, type safety is assured
by phony extra statements which are NOP but would produce warnings
if, for example, the type of the list doesn't correspond to
the element to be stored or retrieved.

This patch and the following ones takes another approach where
all pointer lists are declared *and* defined to essentially the
same structure but kinda parametrized with the type of the pointers
to be stored in the list. This has the advantage that most operations
are done on the real types and thus to eliminate all the type checking
statements and most of the casts (the ones remaining are the ones
needed to interface with the generic type used for the functionalities
implemented as functions).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>